+++ src/main/java/com/be_uterace/repository/EventRepository.java
@@ -10,14 +10,14 @@ import org.springframework.data.repository.query.Param;
 import java.util.Date;
 import java.util.List;
 
-public interface EventRepository extends JpaRepository<Event,Long> {
+public interface EventRepository extends JpaRepository<Event,Integer> {
     @Query("SELECT e FROM Event e WHERE e.status = '1' AND e.endDate > CURRENT_TIMESTAMP")
     List<Event> findAllByStatusAndEndDate();
 
     @Query("SELECT e FROM Event e WHERE e.status = :status AND e.endDate > CURRENT_TIMESTAMP")
     Page<Event> findAllByStatusAndEndDate(String status, Pageable pageable);
 
-    Event findEventByEventId(Long eventId);
+    Event findEventByEventId(Integer eventId);
